chore: move UserMenu component under Header directory
