--- client/src/components/ui/UserMenu/index.ts
@@ -1,1 +1,0 @@
-export * from './user-menu';
